made param_check module and removed from list code
