+++ circularly_linked_list/src/circularly_linked_list.c
@@ -11,20 +11,6 @@
 #include "../include/mod_index.h"
 #include "../include/detection.h"
 
-void param_check (const char * fname, int line_no, int n_args, ...)
-{
-    va_list var_list;
-    va_start(var_list, n_args);
-   
-    for (int i = 0; i < n_args; i++)
-    {
-        char * param = va_arg(var_list, char *);
-        CHECK((i + 1), fname, line_no, param);
-    }
-
-    va_end(var_list);
-}
-
 cll_t * init (uint16_t node_type)
 {
     if (node_type > USR_DEFINE_T)
